refactor: extract connection sorting utilities to shared module (#298)
* Initial plan

* refactor: extract connection sorting utilities to shared module


* fix: use consistent fallback value in JavaScript sorting utility


* fix: remove dependency on undefined sanitizeSortOption in utility script


* fix: use OR operator for better JavaScript compatibility


* fix: use consistent fallback value (999) matching original implementation


* docs: improve documentation for connection sorting utility script


---------
